initial commit to dustin branch
This commit marks the starting point of my branch of changes
I've moved the extension over to webpack so that we could properly disect the initial files present and refactor some of them.
Refactoring will happen naturally as I continue to solve more issues and implement more features, but as of now the end goal is to have relevant functions in different files, and make sure there aren't any functions redifined in each individual file.
We then build the extension up with webpack, which goes through and compiles everything into several source files that the extension can use.
The manifest file already works properly with this new process.
I've fixed the issue of the red tick marker showing when not needed, and i've switched the site management page to use one button to get to instead.